requirements.txt
torch>=1.12.0
torchvision>=0.13.0
opencv-python>=4.5.0
scikit-image>=0.19.0
scikit-fuzzy>=0.4.2
numpy>=1.21.0
pandas>=1.3.0
matplotlib>=3.5.0
flask>=2.0.0
Pillow>=9.0.0
tqdm>=4.64.0